Deleted async/await

+++ LoadFonts_OpenTkOpenGL31CSharp/LoadFonts/Engine/Core/Resources/CharacterInfo.cs
@@ -3,15 +3,15 @@ namespace Engine
 {
     class CharacterInfo
     {
-        public int TexCoordLeft { get; set; }
-        public int TexCoordRight { get; set; }
-        public int TexCoordBottom { get; set; }
-        public int TexCoordTop { get; set; }
+        public float TexCoordLeft { get; set; }
+        public float TexCoordRight { get; set; }
+        public float TexCoordBottom { get; set; }
+        public float TexCoordTop { get; set; }
 
-        public int CharWidth { get; set; }
-        public int CharHeight { get; set; }
-        public int CharWidthOffset { get; set; }
-        public int CharHeightOffset { get; set; }
-        public int CharAspectRatio { get; set; }
+        public float CharWidth { get; set; }
+        public float CharHeight { get; set; }
+        public float CharWidthOffset { get; set; }
+        public float CharHeightOffset { get; set; }
+        public float CharAspectRatio { get; set; }
     }
 }
